+++ Mandatory/includes/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: sbaghdad < sbaghdad@student.1337.ma>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/28 20:41:16 by sbaghdad          #+#    #+#             */
-/*   Updated: 2025/02/28 20:46:19 by sbaghdad         ###   ########.fr       */
+/*   Updated: 2025/03/09 21:56:56 by sbaghdad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "so_long.h"
 
-size_t  ft_strlen(char *s)
+int  ft_strlen(char *s)
 {
     int index;
